Add coverage for Python 3
This commit includes a new environment in tox.ini that generates a
coverage report for Python 3.  This is needed because injector.py (not
yet committed) will have two different implementations of a key function,
depending on the Python version (Python 2 vs. Python 3).  Having the new
"cover3" environment will allow verification that the Python 3 version of
the function has full coverage.