Implement the WSGI application
This commit contains the full implementation of the WSGI application of
Controller.  The entry point is the __call__() method, which constructs a
request, prepares an injector, then calls the dispatch method and handles
the response.  The dispatch method performs the actual resolution step to
determine what to call to handle the request, based on the URL path and
the HTTP method.  The dispatch method also handles some default
functionality, such as the equivalence between HEAD and GET, and a default
implementation of OPTIONS.  Finally, a number of hook methods are defined
to allow consumers to customize various aspects of the behavior of the
micropath framework.