Eliminate merging
This change eliminates the Element.merge() method and all the code related
to it, including the delegation property (which can now be a simple
instance attribute, since the concept of a "master" no longer exists).
Instead, where we would call Element.merge(), we now raise an exception to
diagnose the conflict.