Wrap accessing request attributes
This commit wraps the logic that retrieves attributes from a request with
an exception handler that raises the return value of the new
micropath_request_error() method, with the same debugging behavior as
micropath_server_error().  This ensures that the failure generated in the
case of, for instance, an invalid JSON payload will be, by default, an
HTTPBadRequest (400) error.